backend/pyproject.toml
[tool.poetry]
name = "varuflow-backend"
version = "0.1.0"
description = "Varuflow API — inventory and invoicing for Swedish wholesalers"
authors = ["Varuflow"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.109,<1.0"
uvicorn = {extras = ["standard"], version = ">=0.27"}
sqlalchemy = {extras = ["asyncio"], version = ">=2.0"}
asyncpg = ">=0.29"
alembic = ">=1.13"
pydantic-settings = ">=2.1"
python-dotenv = ">=1.0"
httpx = ">=0.26"
python-jose = {extras = ["cryptography"], version = ">=3.3"}
python-multipart = ">=0.0.9"
reportlab = ">=4.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=4.1"
ruff = ">=0.2"
mypy = ">=1.8"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
